requirements.txt
streamlit
requests
pymongo
numpy
bcrypt
pymodm
bson
transformers
torch
tensorflow
python-dotenv

pymongo[srv]